+++ src/main/kotlin/com/insyncwithfoo/pyright/PyrightInspection.kt
@@ -5,9 +5,11 @@ import com.jetbrains.python.inspections.PyInspection
 
 
 internal class PyrightInspection : PyInspection(), ExternalAnnotatorBatchInspection {
+    
     override fun getShortName() = SHORT_NAME
     
     companion object {
         const val SHORT_NAME = "PyrightInspection"
     }
+    
 }